requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.8.1
    # via
    #   django
    #   django-countries
async-timeout==5.0.1
    # via -r requirements.in
babel==2.17.0
    # via docxcompose3
billiard==4.2.1
    # via celery
boto3==1.34.134
    # via django-storages
botocore==1.34.134
    # via
    #   boto3
    #   s3transfer
celery==5.6.2
    # via -r requirements.in
certifi==2025.1.31
    # via
    #   requests
    #   sentry-sdk
cffi==2.0.0
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.8
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cryptography==44.0.1
    # via
    #   josepy
    #   mozilla-django-oidc
    #   pyopenssl
diff-match-patch==20241021
    # via django-reversion-compare
distlib==0.3.8
    # via virtualenv
dj-importmap==1.0.0
    # via -r requirements.in
django==5.2.11
    # via
    #   -r requirements.in
    #   dj-importmap
    #   django-crispy-forms
    #   django-csp
    #   django-debug-toolbar
    #   django-dirtyfields
    #   django-dsfr
    #   django-filter
    #   django-post-office
    #   django-querysetsequence
    #   django-reversion
    #   django-reversion-compare
    #   django-storages
    #   django-template-partials
    #   django-waffle
    #   mozilla-django-oidc
    #   sentry-sdk
django-countries==8.2.0
    # via -r requirements.in
django-crispy-forms==2.4
    # via django-dsfr
django-csp==4.0
    # via -r requirements.in
django-debug-toolbar==6.2.0
    # via -r requirements.in
django-dirtyfields==1.9.9
    # via -r requirements.in
django-dsfr==3.3.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
django-filter==25.2
    # via -r requirements.in
django-post-office==3.11.0
    # via -r requirements.in
django-querysetsequence==0.18
    # via -r requirements.in
django-reversion==6.1.0
    # via
    #   -r requirements.in
    #   django-reversion-compare
django-reversion-compare==0.19.2
    # via -r requirements.in
django-storages[s3]==1.14.6
    # via -r requirements.in
django-template-partials==25.3
    # via -r requirements.in
django-waffle==5.0.0
    # via -r requirements.in
django-widget-tweaks==1.5.0
    # via django-dsfr
djhtml==3.0.10
    # via -r requirements.in
docxcompose3==1.4.5
    # via -r requirements.in
docxtpl==0.20.2
    # via -r requirements.in
execnet==2.1.1
    # via pytest-xdist
factory-boy==3.3.3
    # via -r requirements.in
faker==37.12.0
    # via
    #   -r requirements.in
    #   factory-boy
filelock==3.13.4
    # via virtualenv
greenlet==3.1.1
    # via playwright
gunicorn==23.0.0
    # via -r requirements.in
identify==2.5.35
    # via pre-commit
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.6
    # via docxtpl
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
josepy==1.14.0
    # via mozilla-django-oidc
kombu==5.6.1
    # via celery
lxml==6.0.2
    # via
    #   docxcompose3
    #   docxtpl
    #   python-docx
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mozilla-django-oidc==4.0.1
    # via -r requirements.in
nh3==0.3.2
    # via django-post-office
nodeenv==1.8.0
    # via pre-commit
packaging==24.0
    # via
    #   django-csp
    #   gunicorn
    #   kombu
    #   pytest
    #   pytest-rerunfailures
platformdirs==4.2.0
    # via virtualenv
playwright==1.58.0
    # via
    #   -r requirements.in
    #   pytest-playwright
pluggy==1.5.0
    # via pytest
pre-commit==4.5.1
    # via -r requirements.in
prompt-toolkit==3.0.50
    # via click-repl
psycopg2-binary==2.9.11
    # via -r requirements.in
pycparser==2.22
    # via cffi
pyee==13.0.0
    # via playwright
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyopenssl==25.0.0
    # via josepy
pytest==8.4.2
    # via
    #   pytest-base-url
    #   pytest-django
    #   pytest-env
    #   pytest-playwright
    #   pytest-rerunfailures
    #   pytest-split
    #   pytest-xdist
pytest-base-url==2.1.0
    # via pytest-playwright
pytest-django==4.11.1
    # via -r requirements.in
pytest-env==1.2.0
    # via -r requirements.in
pytest-playwright==0.7.2
    # via -r requirements.in
pytest-rerunfailures==16.1
    # via -r requirements.in
pytest-split==0.10.0
    # via -r requirements.in
pytest-xdist==3.8.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
python-docx==1.2.0
    # via
    #   docxcompose3
    #   docxtpl
python-magic==0.4.27
    # via -r requirements.in
python-slugify==8.0.4
    # via pytest-playwright
pyyaml==6.0.1
    # via pre-commit
redis==7.1.0
    # via -r requirements.in
requests==2.32.4
    # via
    #   django-dsfr
    #   mozilla-django-oidc
    #   pytest-base-url
    #   requests-mock
requests-mock==1.12.1
    # via -r requirements.in
rich==14.2.0
    # via django-reversion-compare
ruff==0.15.0
    # via -r requirements.in
s3transfer==0.10.2
    # via boto3
sentry-sdk[django]==2.52.0
    # via -r requirements.in
six==1.16.0
    # via
    #   docxcompose3
    #   python-dateutil
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.11.0
    # via
    #   django-countries
    #   pyee
    #   python-docx
tzdata==2025.2
    # via
    #   faker
    #   kombu
tzlocal==5.3.1
    # via celery
urllib3==2.6.3
    # via
    #   botocore
    #   requests
    #   sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.29.1
    # via pre-commit
wcwidth==0.2.13
    # via prompt-toolkit
whitenoise==6.11.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
